+++ plugins/connection/network_cli_test.py
@@ -6,30 +6,27 @@ from __future__ import absolute_import, division, print_function
 
 __metaclass__ = type
 
-import json
 import os
-import re
 
-from ansible.errors import AnsibleError
 from ansible_collections.ansible.netcommon.plugins.connection.network_cli import (
     DOCUMENTATION,
     Connection as RealConnection,
 )
+from ansible_collections.ansible.netcommon.plugins.plugin_utils.test_connection import (
+    TestConnection,
+    record_response,
+    compare_response,
+)
 
 
-class Connection(RealConnection):
-    def __init__(self, play_context, new_stdin, *args, **kwargs):
-        super(Connection, self).__init__(
-            play_context, new_stdin, *args, **kwargs
-        )
-        # Track the send sequence for testing, increment before send
-        self._send_sequence = -1
+DOCUMENTATION += """
+extends_documentation_fragment:
+- ansible.netcommon.connection_persistent
+- ansible.netcommon.test_parameters
+"""
 
-    def _connect(self):
-        if self.get_option("test_parameters").get("mode") == "playback":
-            return
-        return super(Connection, self)._connect()
 
+class Connection(TestConnection, RealConnection):
     def send(
         self,
         command,
@@ -42,7 +39,7 @@ class Connection(RealConnection):
         strip_prompt=True,
     ):
         # Check for testing
-        if self.get_option("test_parameters").get("test_mode") == "playback":
+        if self.get_option("test_parameters").get("mode") == "playback":
             return self._send_playback(command)
         response = super(Connection, self).send(
             command,
@@ -67,73 +64,8 @@ class Connection(RealConnection):
         )
 
         if test_parameters["mode"] == "record":
-            os.makedirs(test_parameters["fixture_directory"], exist_ok=True)
-
-            with open(fixture_file, "w") as f:
-                json.dump(
-                    {"command": command.decode("utf-8"), "response": response},
-                    f,
-                )
-                f.write("\n")
+            record_response(command, response, fixture_file, test_parameters)
         elif test_parameters["mode"] == "compare":
-            with open(fixture_file, "r") as f:
-                fixture = json.load(f)
-            if fixture["command"] != command.decode("utf-8"):
-                raise AssertionError(
-                    (
-                        "Command sent to device does not match the recorded command",
-                        fixture["command"],
-                        command.decode("utf-8"),
-                    ),
-                )
-            fixture_lines = fixture["response"].splitlines()
-            response_lines = response.splitlines()
-            if len(fixture_lines) != len(response_lines):
-                raise AssertionError(
-                    (
-                        "Response length does not match the recorded response",
-                        len(fixture["response"]),
-                        len(response),
-                    ),
-                )
-            for idx, line in enumerate(fixture_lines):
-                if line != response_lines[idx]:
-                    if any(
-                        re.match(re_exempt, line)
-                        for re_exempt in test_parameters["exempted"]
-                    ):
-                        continue
-                    raise AssertionError(
-                        (
-                            "Response line does not match the recorded response line",
-                            line,
-                            response[idx],
-                        ),
-                    )
+            compare_response(command, response, fixture_file, test_parameters)
+
         return response
-
-    def _send_playback(self, command):
-        """Send the fixture response rather than the actual command."""
-
-        test_parameters = self.get_option("test_parameters")
-
-        self._send_sequence += 1
-
-        fixture_file = os.path.join(
-            test_parameters["fixture_directory"],
-            "%s.json" % self._send_sequence,
-        )
-        if not os.path.exists(fixture_file):
-            raise AnsibleError(
-                "Fixture file %s does not exist." % fixture_file
-            )
-
-        with open(fixture_file, "r") as f:
-            fixture = json.load(f)
-
-        if fixture["command"] != command:
-            raise AssertionError(
-                "Fixture command %s does not match command %s."
-                % (fixture["command"], command)
-            )
-        return fixture["response"]
